refactor: remove extra interface for service
