+++ src/main/java/org/tbank/hw8/service/JwtService.java
@@ -1,4 +1,4 @@
-package org.tbank.hw8.service.impl;
+package org.tbank.hw8.service;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
